Make L0_backend_python Windows friendly

+++ qa/L0_backend_python/ensemble/test.sh
@@ -26,14 +26,17 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 CLIENT_LOG="./ensemble_client.log"
-EXPECTED_NUM_TESTS="2"
 TEST_RESULT_FILE='test_results.txt'
 source ../common.sh
 source ../../common/util.sh
 
-TRITON_DIR=${TRITON_DIR:="/opt/tritonserver"}
-SERVER=${TRITON_DIR}/bin/tritonserver
-BACKEND_DIR=${TRITON_DIR}/backends
+# FIXME: Until Windows supports GPU tensors, only test CPU scenarios
+if [[ ${WINDOWS} == 1 ]]; then
+    EXPECTED_NUM_TESTS="1"
+else
+    EXPECTED_NUM_TESTS="2"
+fi
+
 SERVER_ARGS="--model-repository=`pwd`/models --backend-directory=${BACKEND_DIR} --log-verbose=1"
 SERVER_LOG="./ensemble_server.log"
 
@@ -52,18 +55,21 @@ mkdir -p models/add_sub_2/1/
 cp ../../python_models/add_sub/config.pbtxt ./models/add_sub_2/
 cp ../../python_models/add_sub/model.py ./models/add_sub_2/1/
 
-# Ensemble GPU Model
-mkdir -p models/ensemble_gpu/1/
-cp ../../python_models/ensemble_gpu/config.pbtxt ./models/ensemble_gpu
-cp -r ${DATADIR}/qa_model_repository/libtorch_float32_float32_float32/ ./models
-(cd models/libtorch_float32_float32_float32 && \
-          echo "instance_group [ { kind: KIND_GPU }]" >> config.pbtxt)
-(cd models/libtorch_float32_float32_float32 && \
-          sed -i "s/^max_batch_size:.*/max_batch_size: 0/" config.pbtxt)
-(cd models/libtorch_float32_float32_float32 && \
-          sed -i "s/^version_policy:.*//" config.pbtxt)
-rm -rf models/libtorch_float32_float32_float32/2
-rm -rf models/libtorch_float32_float32_float32/3
+# FIXME: Until Windows supports GPU tensors, only test CPU scenarios
+if [[ ${WINDOWS} == 0 ]]; then
+    # Ensemble GPU Model
+    mkdir -p models/ensemble_gpu/1/
+    cp ../../python_models/ensemble_gpu/config.pbtxt ./models/ensemble_gpu
+    cp -r ${DATADIR}/qa_model_repository/libtorch_float32_float32_float32/ ./models
+    (cd models/libtorch_float32_float32_float32 && \
+            echo "instance_group [ { kind: KIND_GPU }]" >> config.pbtxt)
+    (cd models/libtorch_float32_float32_float32 && \
+            sed -i "s/^max_batch_size:.*/max_batch_size: 0/" config.pbtxt)
+    (cd models/libtorch_float32_float32_float32 && \
+            sed -i "s/^version_policy:.*//" config.pbtxt)
+    rm -rf models/libtorch_float32_float32_float32/2
+    rm -rf models/libtorch_float32_float32_float32/3
+fi
 
 prev_num_pages=`get_shm_pages`
 
@@ -75,7 +81,13 @@ if [ "$SERVER_PID" == "0" ]; then
 fi
 
 set +e
-python3 ensemble_test.py 2>&1 > $CLIENT_LOG
+
+# FIXME: Until Windows supports GPU tensors, only test CPU scenarios
+if [[ ${WINDOWS} == 0 ]]; then
+    python3 ensemble_test.py 2>&1 > $CLIENT_LOG
+else
+    python3 ensemble_test.py EnsembleTest.test_ensemble
+fi
 
 if [ $? -ne 0 ]; then
     echo -e "\n***\n*** ensemble_test.py FAILED. \n***"
